Several bug fixes.

+++ cosmos/models/mobility/src/main/java/es/tid/cosmos/mobility/labeljoining/ClusterAggBtsClusterReducer.java
@@ -5,7 +5,6 @@ import java.util.LinkedList;
 import java.util.List;
 
 import com.twitter.elephantbird.mapreduce.io.ProtobufWritable;
-import org.apache.hadoop.io.IntWritable;
 import org.apache.hadoop.io.LongWritable;
 import org.apache.hadoop.mapreduce.Reducer;
 
@@ -19,7 +18,7 @@ import es.tid.cosmos.mobility.data.PoiUtil;
  * @author dmicol
  */
 public class ClusterAggBtsClusterReducer extends Reducer<LongWritable,
-        ProtobufWritable<MobData>, IntWritable, ProtobufWritable<Poi>> {
+        ProtobufWritable<MobData>, LongWritable, ProtobufWritable<Poi>> {
     @Override
     protected void reduce(LongWritable key,
             Iterable<ProtobufWritable<MobData>> values, Context context)
@@ -48,7 +47,7 @@ public class ClusterAggBtsClusterReducer extends Reducer<LongWritable,
                 outputPoi.setLabelgroupbts(cluster.getLabelgroup());
                 outputPoi.setConfidentbts(cluster.getConfident());
                 outputPoi.setDistancebts(cluster.getDistance());
-                context.write(new IntWritable(poi.getBts()),
+                context.write(new LongWritable(poi.getBts()),
                               PoiUtil.wrap(outputPoi.build()));
             }
         }
